Move dao/entity/repo to queryChannels package
